refactor ui app trace detail loader
